changing float conversion

+++ go/armstrong-numbers/armstrong_numbers.go
@@ -10,14 +10,14 @@ import (
 func IsNumber(number int) bool {
 	s := strings.Split(strconv.Itoa(number), "")
 
-	factor := len(s)
+	factor := float64(len(s))
 	var total float64
 	for _, v := range s {
 		n, err := strconv.ParseFloat(v, 64)
 		if err != nil {
 			return false
 		}
-		total += math.Pow(n, float64(factor))
+		total += math.Pow(n, factor)
 	}
 	return number == int(total)
 }
